UL: Driver inside Bidirectional
